backend/requirements.txt
# Web Framework
fastapi==0.115.12
uvicorn==0.34.2
starlette==0.46.2

# WebSocket Support
websockets==15.0.1

# Database
mysql-connector-python==9.3.0

# Authentication
bcrypt==3.2.0

# Environment Variables
python-dotenv==1.1.1

# Data Validation
pydantic==2.11.4
pydantic-settings==2.10.1
pydantic_core==2.33.2

# Image Processing & AI
ultralytics==8.3.126
opencv-python==4.11.0.86
pillow==11.0.0
torch==2.7.0+cu118
torchvision==0.22.0+cu118

# Utilities
numpy==2.1.2
python-multipart==0.0.20

# HTTP Client
requests==2.25.1

# Other Dependencies
annotated-types==0.7.0
anyio==4.9.0
certifi==2020.6.20
click==8.0.3
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
psutil==7.0.0
py-cpuinfo==9.0.0
PyYAML==5.4.1
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
